Update CommandParser.cs
in line 82 you can see a switch statement being used which allows the different commands to be separated, all the different commands check if the right amount of parts are used. for eg; rectangle has 3 parts and circle has 2 parts. if the right amount of parts aren't used the command will not run and come up false.